requirements.txt
alembic>=0.4.1
argparse
Babel>=1.3
cryptography>=0.5
eventlet>=0.13.0
iso8601>=0.1.9
jsonschema>=1.3.0,!=1.4.0
kombu>=2.4.8
netaddr>=0.7.6
oslo.config>=1.2.0
oslo.messaging>=1.3.0a4
Paste
PasteDeploy>=1.5.0
pbr>=0.6,!=0.7,<1.0
pecan>=0.6.0
pycrypto>=2.6.1
pysqlite
python-keystoneclient>=0.9.0
six>=1.7.0
SQLAlchemy>=0.7.8,<=0.9.99
stevedore>=0.12
WebOb>=1.2.3,<1.3
wsgiref>=0.1.2
thrift>=0.9.1

# TODO: Get this working again...PyKCS11>=1.2.4
